requirements.txt
pandas
numpy
mysql-connector-python
tensorflow
scikit-learn
matplotlib
requests
python-dotenv
